Refactor ETL Modules
